Output WARNING/ERROR if commits are unpushed
If the commit toni is about to send to Applikatoni has not been pushed
the deployment would fail. In this case we output an ERROR and exit with
a non-zero exit code.

If the commit has been pushed to the remote and other changes haven't
been pushed to the remote, that's maybe not what the user wants. In this
case output a WARNING but keep on deploying.